test_system_health.py
#!/usr/bin/env python3
"""
System health check script
Tests all critical dependencies and functionality
"""
import sys
import importlib
from pathlib import Path

def test_basic_imports():
    """Test basic Python dependencies"""
    results = {}
    
    # Critical packages
    critical_packages = {
        'torch': 'PyTorch',
        'numpy': 'NumPy',
        'pandas': 'Pandas',
        'fastapi': 'FastAPI',
        'pdfplumber': 'PDF Processing',
        'pytesseract': 'OCR',
        'jieba': 'Chinese NLP',
        'networkx': 'Graph Processing',
        'sqlite3': 'SQLite Database',
        'json': 'JSON Processing',
        'pathlib': 'Path Handling',
    }
    
    for package, description in critical_packages.items():
        try:
            importlib.import_module(package)
            results[description] = "✓ Available"
        except ImportError as e:
            results[description] = f"✗ Failed: {e}"
    
    return results

def test_ocr_functionality():
    """Test OCR capabilities"""
    try:
        import pytesseract
        import tempfile
        from PIL import Image, ImageDraw
        
        # Create test image
        img = Image.new('RGB', (200, 50), color='white')
        d = ImageDraw.Draw(img)
        d.text((10, 10), 'Test Text', fill='black')
        
        with tempfile.NamedTemporaryFile(suffix='.png', delete=False) as tmp:
            img.save(tmp.name)
            text = pytesseract.image_to_string(tmp.name).strip()
            
        return "✓ OCR working" if text else "✓ OCR available (no text detected)"
        
    except Exception as e:
        return f"✗ OCR failed: {e}"

def test_pdf_processing():
    """Test PDF processing capabilities"""
    try:
        import pdfplumber
        return "✓ PDF processing available"
    except Exception as e:
        return f"✗ PDF processing failed: {e}"

def test_chinese_nlp():
    """Test Chinese NLP capabilities"""
    try:
        import jieba
        words = list(jieba.cut("易经是古代哲学经典"))
        return f"✓ Chinese NLP working (segmented {len(words)} words)"
    except Exception as e:
        return f"✗ Chinese NLP failed: {e}"

def test_database_connectivity():
    """Test SQLite database functionality"""
    try:
        import sqlite3
        import tempfile
        
        with tempfile.NamedTemporaryFile(suffix='.db', delete=False) as tmp:
            conn = sqlite3.connect(tmp.name)
            cursor = conn.cursor()
            cursor.execute("CREATE TABLE test (id INTEGER PRIMARY KEY)")
            cursor.execute("INSERT INTO test VALUES (1)")
            conn.commit()
            conn.close()
            
        return "✓ SQLite database working"
    except Exception as e:
        return f"✗ Database test failed: {e}"

def check_environment():
    """Check environment and system information"""
    info = {
        'Python Version': f"{sys.version_info.major}.{sys.version_info.minor}.{sys.version_info.micro}",
        'Python Path': sys.executable,
        'Working Directory': str(Path.cwd()),
        'Platform': sys.platform,
    }
    return info

def main():
    """Run comprehensive system health check"""
    print("=" * 60)
    print("SYSTEM HEALTH CHECK - 易学PDF处理系统")
    print("=" * 60)
    
    # Environment info
    print("\n🔍 ENVIRONMENT INFO:")
    env_info = check_environment()
    for key, value in env_info.items():
        print(f"   {key}: {value}")
    
    # Basic imports
    print("\n📦 PACKAGE DEPENDENCIES:")
    basic_results = test_basic_imports()
    for description, result in basic_results.items():
        print(f"   {description}: {result}")
    
    # Functionality tests
    print("\n🧪 FUNCTIONALITY TESTS:")
    print(f"   OCR System: {test_ocr_functionality()}")
    print(f"   PDF Processing: {test_pdf_processing()}")
    print(f"   Chinese NLP: {test_chinese_nlp()}")
    print(f"   Database: {test_database_connectivity()}")
    
    # Docker status
    print("\n🐳 DOCKER STATUS:")
    try:
        import subprocess
        result = subprocess.run(['docker', '--version'], capture_output=True, text=True)
        if result.returncode == 0:
            print(f"   Docker: ✓ Available - {result.stdout.strip()}")
        else:
            print("   Docker: ✗ Docker Desktop WSL integration needed")
    except FileNotFoundError:
        print("   Docker: ⚠ Not found - Enable Docker Desktop WSL2 integration")
    
    # File structure check
    print("\n📁 KEY FILE STRUCTURE:")
    key_paths = [
        'requirements.txt',
        'docker-compose.yml',
        'api/main.py',
        'extract_pdfs.py',
        'knowledge_graph/graph_builder.py'
    ]
    
    for path in key_paths:
        if Path(path).exists():
            print(f"   {path}: ✓ Present")
        else:
            print(f"   {path}: ✗ Missing")
    
    print("\n" + "=" * 60)
    print("SYSTEM STATUS: READY FOR DEVELOPMENT")
    print("=" * 60)
    
    print("\n📝 NEXT STEPS:")
    print("   1. Enable Docker Desktop WSL2 integration for containerized deployment")
    print("   2. Use individual scripts for testing specific functionality")
    print("   3. Run production pipeline with: python production_extract.py")
    print("   4. Start API server with: uvicorn api.main:app --reload")

if __name__ == "__main__":
    main()